requirements.txt
streamlit==1.33.0
ultralytics==8.3.94
ultralytics-thop==2.0.14
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
pandas==2.2.2
PyYAML==6.0.1
scipy==1.13.0
torch==2.1.0
torchvision==0.16.0